course section added

+++ src/app/services/destination.service.ts
@@ -1,8 +1,19 @@
 import { Destination } from '../models/destination.model';
 import { IDestination } from '../interfaces/destination.interface';
+import { cloudinary } from '../utils/cloudinary';
 
 const createDestination = async (data: IDestination) => {
-  return await Destination.create(data);
+  try {
+    return await Destination.create(data);
+  } catch (error) {
+    if (data.photo) {
+      const publicId = data.photo.split('/').pop()?.split('.')[0];
+      if (publicId) {
+        await cloudinary.uploader.destroy(publicId);
+      }
+    }
+    throw error;
+  }
 };
 
 const getDestinations = async (filters = {}, sort = '-createdAt') => {
@@ -14,10 +25,33 @@ const getDestinationById = async (id: string) => {
 };
 
 const updateDestination = async (id: string, data: Partial<IDestination>) => {
-  return await Destination.findByIdAndUpdate(id, data, { new: true });
+  try {
+    return await Destination.findByIdAndUpdate(id, data, {
+      new: true,
+      runValidators: true,
+    });
+  } catch (error) {
+    if (data.photo) {
+      const publicId = data.photo.split('/').pop()?.split('.')[0];
+      if (publicId) {
+        await cloudinary.uploader.destroy(publicId);
+      }
+    }
+    throw error;
+  }
 };
 
 const deleteDestination = async (id: string) => {
+  const existing = await Destination.findById(id);
+  if (!existing) return null;
+
+  if (existing.photo) {
+    const publicId = existing.photo.split('/').pop()?.split('.')[0];
+    if (publicId) {
+      await cloudinary.uploader.destroy(publicId);
+    }
+  }
+
   return await Destination.findByIdAndDelete(id);
 };
